Updated to to what may be a newer jquery.history.js.
"Fixed" comic load.

My guess is this is not exactly the desired way to do this, but I don't know enough about Ghost to know how to force it to pull out post X and repaint without doing a browser reload.

This makes it functional, even if not the desired end goal.  Note the only change is the commented out line.